stage 2 done

+++ config/logging.php
@@ -122,7 +122,7 @@ return [
             'driver' => 'custom',
             'via' => \App\Logging\Telegram\TelegramLoggerFactory::class,
             'level' => env('LOG_LEVEL', 'debug'),
-            'chat_id' => (int) env('LOGGER_TELEGRAM_CHAT_ID', ''),
+            'chat_id' => env('LOGGER_TELEGRAM_CHAT_ID', ''),
             'token' => env('LOGGER_TELEGRAM_TOKEN', ''),
         ],
     ],
